Stop relying on customer fillable content to create s3 keys
Using the filename on the s3 key is not a good idea as these can contain lots of weird characters and open ways for security breaches.
The UUID should be enough to guarantee uniqueness on the keys and the fieldName should be enough to at least point to a direction on what this resource is.

This is also not breaking as the href field of old webresources should still work fine but the new ones shall not use possibly skewed data.

Change-type: patch